--- src/setup_node_env/harden.js
@@ -1,49 +1,0 @@
-/*
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-var hook = require('require-in-the-middle');
-var createLodashTemplateProxy = require('./patches/lodash_template');
-var createLodashFpTemplateProxy = require('./patches/lodash_fp_template');
-
-hook(['child_process'], function (exports, name) {
-  return require(`./patches/${name}`)(exports); // eslint-disable-line import/no-dynamic-require
-});
-
-hook(['lodash'], function (exports) {
-  return {
-    ...exports,
-    template: createLodashTemplateProxy(exports.template),
-  };
-});
-
-hook(['lodash/template'], function (exports) {
-  return createLodashTemplateProxy(exports);
-});
-
-// the lodash_fp patches relies on lodash already being patched... the order matters here
-hook(['lodash/fp'], function (exports) {
-  return {
-    ...exports,
-    template: createLodashFpTemplateProxy(exports.template),
-  };
-});
-
-hook(['lodash/fp/template'], function (exports) {
-  return createLodashFpTemplateProxy(exports);
-});
