+++ src/setup_node_env/harden/index.js
@@ -17,16 +17,5 @@
  * under the License.
  */
 
-module.exports = function (template) {
-  var _ = require('lodash');
-
-  // per https://github.com/lodash/lodash/wiki/FP-Guide
-  // > Iteratee arguments are capped to avoid gotchas with variadic iteratees.
-  // this means that we can't specify the options in the second argument to fp.template because it's ignored.
-  // Instead, we're going to use the non-FP _.template with only the first argument which has already been patched
-  return new Proxy(template, {
-    apply: function (target, thisArg, args) {
-      return _.template(args[0]);
-    },
-  });
-};
+require('./child_process');
+require('./lodash_template');
